# docs/swagger.yaml
definitions:
...
  github_com_mr-isik_gatling-backend_internal_domain.Anomaly:
    properties:
      message:
        type: string
      severity:
        description: low, medium, high
        type: string
      timestamp:
        type: string
      type:
        description: latency_spike, error_surge, throughput_drop
        type: string
    type: object
  github_com_mr-isik_gatling-backend_internal_domain.Check:
    properties:
      operator:
        description: eq, gt, contains
        type: string
      type:
        description: status, body_contains, json_path
        type: string
      value:
        type: string
    type: object
...
  github_com_mr-isik_gatling-backend_internal_domain.Report:
    properties:
      ai_recommendations:
        items:
          type: string
        type: array
      ai_summary:
        type: string
      anomalies:
        items:
          $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Anomaly'
        type: array
      created_at:
        type: string
      id:
        type: string
      run_id:
        type: string
      summary:
        $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Summary'
    type: object
...
  github_com_mr-isik_gatling-backend_internal_domain.Scenario:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      is_ai_generated:
        type: boolean
      name:
        type: string
      project_id:
        type: string
      steps:
        items:
          $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Step'
        type: array
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  github_com_mr-isik_gatling-backend_internal_domain.Step:
    properties:
      body:
        type: string
      checks:
        items:
          $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Check'
        type: array
      headers:
        additionalProperties:
          type: string
        type: object
      method:
        type: string
      name:
        type: string
      order:
        type: integer
      think_time_ms:
        type: integer
      url:
        type: string
    type: object
  github_com_mr-isik_gatling-backend_internal_domain.Summary:
    properties:
      avg_latency:
        type: number
      duration_ms:
        type: integer
      error_count:
        type: integer
      max_vus:
        type: integer
      p95_latency:
        type: number
      p99_latency:
        type: number
      success_count:
        type: integer
      total_requests:
        type: integer
    type: object
...
  internal_api_handler.cloneRequest:
    properties:
      name:
        type: string
    type: object
...
  internal_api_handler.exportRequest:
    properties:
      format:
        type: string
    type: object
  internal_api_handler.generateRequest:
    properties:
      prompt:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI-Powered Load Testing Tool Backend API.
  termsOfService: http://swagger.io/terms/
  title: Gatling-Backend API
  version: "1.0"
paths:
...
  /v1/reports/{id}:
    get:
      description: Gets a specific test report by ID.
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Report'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Report
      tags:
      - Report
  /v1/reports/{id}/ai-summary:
    get:
      description: Gets an AI generated summary for a specific test report.
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Report'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Summary
      tags:
      - Report
  /v1/reports/{id}/export:
    post:
      consumes:
      - application/json
      description: Exports a report in the specified format (e.g. PDF, CSV).
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Export Format
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.exportRequest'
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Export Report
      tags:
      - Report
  /v1/reports/compare:
    get:
      description: Compares two test runs (baseline vs current).
      parameters:
      - description: Baseline Run ID
        in: query
        name: run1
        required: true
        type: string
      - description: Current Run ID
        in: query
        name: run2
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Compare Reports
      tags:
      - Report
...
  /v1/scenarios:
    get:
      description: Lists all test scenarios for a given project.
      parameters:
      - description: Project ID
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Scenario'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Scenarios
      tags:
      - Scenario
    post:
      consumes:
      - application/json
      description: Creates a new test scenario.
      parameters:
      - description: Scenario Data
        in: body
        name: scenario
        required: true
        schema:
          $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Scenario'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Scenario'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Scenario
      tags:
      - Scenario
  /v1/scenarios/{id}:
    delete:
      description: Deletes a scenario from the system by ID.
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete Scenario
      tags:
      - Scenario
    get:
      description: Gets detailed information for a specific scenario by ID.
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Scenario'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Scenario
      tags:
      - Scenario
    put:
      consumes:
      - application/json
      description: Updates the data of an existing scenario by ID.
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Scenario Data
        in: body
        name: scenario
        required: true
        schema:
          $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Scenario'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Scenario'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Scenario
      tags:
      - Scenario
  /v1/scenarios/{id}/clone:
    post:
      consumes:
      - application/json
      description: Clones an existing scenario with a new name.
      parameters:
      - description: Scenario ID to clone
        in: path
        name: id
        required: true
        type: string
      - description: New Scenario Name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.cloneRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Scenario'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Clone Scenario
      tags:
      - Scenario
  /v1/scenarios/generate:
    post:
      consumes:
      - application/json
      description: Generates a test scenario using AI based on a prompt.
      parameters:
      - description: AI Prompt
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.generateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Scenario'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Generate Scenario via AI
      tags:
      - Scenario
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"